# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OCPP Central System API",
        "description": "REST API for managing OCPP 1.6 charge points and transactions",
        "contact": {
            "name": "Texnouz",
            "email": "support@texnouz.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/tariffs": {
            "get": {
                "tags": [
                    "tariffs"
                ],
                "summary": "List all tariffs",
                "operationId": "list_tariffs",
                "responses": {
                    "200": {
                        "description": "List of tariffs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Create a new tariff",
                "operationId": "create_tariff",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTariffRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Tariff created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/tariffs/default": {
            "get": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Get the default tariff",
                "operationId": "get_default_tariff",
                "responses": {
                    "200": {
                        "description": "Default tariff",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No default tariff found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/tariffs/preview-cost": {
            "post": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Preview cost calculation",
                "operationId": "preview_cost",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CostPreviewRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Cost breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tariff not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
        "/api/v1/tariffs/{id}": {
            "get": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Get a tariff by ID",
                "operationId": "get_tariff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tariff ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tariff details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tariff not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Update a tariff",
                "operationId": "update_tariff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tariff ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTariffRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Tariff updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tariff not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tariffs"
                ],
                "summary": "Delete a tariff",
                "operationId": "delete_tariff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tariff ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tariff deleted"
                    },
                    "404": {
                        "description": "Tariff not found"
                    }
                },
                "security": [
                    {
                        "bearer_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/T"
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "CostPreviewRequest": {
                "type": "object",
                "description": "Cost preview request DTO",
                "required": [
                    "energy_wh",
                    "duration_seconds"
                ],
                "properties": {
                    "duration_seconds": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Duration in seconds"
                    },
                    "energy_wh": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Energy consumed in Wh"
                    },
                    "tariff_id": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Tariff ID to use (optional, uses default if not specified)",
                        "nullable": true
                    }
                }
            },
...
            "CreateTariffRequest": {
                "type": "object",
                "description": "Create tariff request DTO",
                "required": [
                    "name",
                    "tariff_type",
                    "price_per_kwh",
                    "price_per_minute",
                    "session_fee",
                    "currency"
                ],
                "properties": {
                    "currency": {
                        "type": "string",
                        "description": "Currency code (ISO 4217)"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "is_default": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "max_fee": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Maximum fee (0 = no limit)",
                        "nullable": true
                    },
                    "min_fee": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Minimum fee",
                        "nullable": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "price_per_kwh": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Price per kWh in smallest currency unit (e.g., cents)"
                    },
                    "price_per_minute": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Price per minute in smallest currency unit"
                    },
                    "session_fee": {
                        "type": "integer",
                        "format": "int32",
                        "description": "Session start fee in smallest currency unit"
                    },
                    "tariff_type": {
                        "type": "string",
                        "description": "Tariff type: PerKwh, PerMinute, PerSession, Combined"
                    },
                    "valid_from": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                }
            },
...
            "UpdateTariffRequest": {
                "type": "object",
                "description": "Update tariff request DTO",
                "properties": {
                    "currency": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "is_default": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "max_fee": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "min_fee": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "price_per_kwh": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "price_per_minute": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "session_fee": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "tariff_type": {
                        "type": "string",
                        "nullable": true
                    },
                    "valid_from": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "valid_until": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                }
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            },
            "bearer_auth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token"
            }
        }
    },
...
}